Bluntly destroy all of the audio controls when destroying.
Apparently something went wrong when destroying individual widgets. But why
take all the trouble if there's an easy way to simply destroy all of it. I
checked the memory usage. Before this commit there was a growing RSS memory
usage when constantly reversing, but after this commit, that is gone and I see
no change in memory usage. So I guess this also fixes some latent memory leak.

Fixes #35.