Added fullspeedwhenloading control in Misc controls tab
At the same time, fixed the layout a bit, regarding the Fast Fwd button
position. It should have been at the top, but it wasn't. Now it is and the new
button for fullspeedwhenloading is as well.
While struggling with this, also found some left over sizers that weren't used
and removed them.