build: Remove CPU detection
Catapult doesn't have any low-level code that cares about which CPU
it is running on. The only thing the CPU name was used for is the
output dir, but I don't think a lot of people will be cross-compiling
Catapult, so that alone is not worth all this extra code.