Let "make install" do the right thing
This lets make install do the right thing. Instead of one 16x16 xpm
icon, you now get a nice desktop icon, just like for openMSX itself.

Also updated the .desktop file, this matches what is used for the
flatpak. And extract the PNG files from the ICO file for use on
Linux.

And change the .desktop file to reverse DNS notation as the
freedesktop standard nowadays indicates you should use. This also
matches what is mandated by flatpak and will all help to simplify
packaging the next openMSX, not only for flatpak, but also for RPM.