Fix check for supported warning flag
Clang classifies the passing of an unsupported warning flag as
a warning in itself, so it won't return a non-zero exit code unless
we also pass `-Werror`.

GCC will default to interpreting stdin as C, even if it's called
from via the `g++` executable name, so add `-x c++` to force it
to treat stdin as C++. This is important since some warnings are
C++-only.