Create Visual Studio GitHub Actions build
This is just to get started with automatically building Catapult. For Windows, this is the most complex, so that's why I started there. This is open for expansion to multiple platforms.

As the old libxml2-v2.8.0 didn't build with a modern Visual Studio as is available on the runners, I upgraded it at first to 2.9.3 (that works) and later I just updated both libxml2 and wxWidgets to the latest stable releases. That was tested by GrayFoxNL on Discord, and was running fine.

This autobuild also makes use of the caching for the 3rdparty stuff that is downloaded and built.

The build/version.py file that comes with Catapult was an old duplicate of the one in openMSX. I synced it with the current openMSX version, to get similar output. This is useful to give meaningful names to the files.

The organization of files in the zip file is mimicked from the zip delivery according to the package-windows scripting in the openMSX repo.

Thanks to Vampier for passing me the script he uses to build Catapult locally. That is the basis for this workflow.